# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ECOnexão API
  description: API de domínio do aplicativo ECOnexão para turismo sustentável e rotas
    ecológicas.
  version: 1.0.0
paths:
...
  /api/v1/admin/media/process:
    post:
      tags:
      - Admin Media
      summary: Process Editorial Media
      operationId: process_editorial_media_api_v1_admin_media_process_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_process_editorial_media_api_v1_admin_media_process_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditorialMediaEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/media/cleanup/recover:
    post:
      tags:
      - Admin Media
      summary: Recover Media Cleanup
      operationId: recover_media_cleanup_api_v1_admin_media_cleanup_recover_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanupRecoveryRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanupRecoveryEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/actors/{actor_id}:
    get:
      tags:
      - Territorial - Actors
      summary: Detalhes de um ator
      description: Retorna informações completas de um ator ou estabelecimento específico.
      operationId: get_actor_detail_api_v1_actors__actor_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: actor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Actor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDetailEnvelope'
        '404':
          description: Ator não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/actors/{actor_id}/google-photo:
    get:
      tags:
      - Territorial - Actors
      summary: Foto Google temporária do ator
      description: Emite um grant opaco a partir de Place Details recente; não expõe URLs Google.
      operationId: get_actor_google_photo_api_v1_actors__actor_id__google_photo_get
      parameters:
      - name: actor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Actor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GooglePhotoMetadataEnvelope'
        '404':
          description: Foto indisponível.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Foto temporariamente indisponível.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/places/photos/{token}:
    get:
      tags:
      - Google Place Photos
      summary: Get Place Photo
      operationId: get_place_photo_api_v1_places_photos__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      - name: maxHeightPx
        in: query
        required: false
        schema:
          type: integer
          maximum: 4800
          minimum: 1
          default: 600
          title: Maxheightpx
      - name: maxWidthPx
        in: query
        required: false
        schema:
          type: integer
          maximum: 4800
          minimum: 1
          default: 800
          title: Maxwidthpx
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Foto indisponível
        '410':
          description: Foto expirada
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ActorCategorySchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        slug:
          type: string
          title: Slug
        label:
          type: string
          title: Label
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        sort_order:
          type: integer
          title: Sort Order
          default: 0
      type: object
      required:
      - id
      - label
      - slug
      - sort_order
      title: ActorCategorySchema
    ActorDetailEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/ActorDetailSchema'
      type: object
      required:
      - data
      title: ActorDetailEnvelope
    ActorDetailSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        slug:
          type: string
          title: Slug
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        category:
          $ref: '#/components/schemas/ActorCategorySchema'
        sub_category:
          anyOf:
          - type: string
          - type: 'null'
          title: Sub Category
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        city:
          anyOf:
          - type: string
          - type: 'null'
          title: City
        state_code:
          anyOf:
          - type: string
          - type: 'null'
          title: State Code
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        instagram:
          anyOf:
          - type: string
          - type: 'null'
          title: Instagram
        website:
          anyOf:
          - type: string
          - type: 'null'
          title: Website
        opening_hours:
          additionalProperties: true
          type: object
          title: Opening Hours
        payment_methods:
          items: {}
          type: array
          title: Payment Methods
        green_badge_status:
          type: string
          title: Green Badge Status
          default: none
        verification_status:
          type: string
          title: Verification Status
          default: unverified
        google_place_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Google Place Id
        google_rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Google Rating
        google_review_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Google Review Count
        cover_image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Cover Image Url
        cover_media:
          anyOf:
          - $ref: '#/components/schemas/ResolvedMediaItemSchema'
          - type: 'null'
        is_favorite:
          type: boolean
          title: Is Favorite
          default: false
        gallery:
          items:
            $ref: '#/components/schemas/ResolvedMediaItemSchema'
          type: array
          title: Gallery
        accessibility_features:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Accessibility Features
      type: object
      required:
      - accessibility_features
      - category
      - green_badge_status
      - id
      - name
      - opening_hours
      - payment_methods
      - slug
      - verification_status
      title: ActorDetailSchema
...
    Body_process_editorial_media_api_v1_admin_media_process_post:
      properties:
        owner_type:
          type: string
          enum:
          - route
          - origin
          - actor
          title: Owner Type
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        alt_text:
          type: string
          maxLength: 500
          minLength: 1
          title: Alt Text
        credit:
          type: string
          maxLength: 500
          minLength: 1
          title: Credit
        license_code:
          type: string
          enum:
          - CC-BY-4.0
          - SEMTUR_INSTITUTIONAL
          - PROPRIETARY
          title: License Code
        image:
          type: string
          format: binary
          title: Image
      type: object
      required:
      - owner_type
      - owner_id
      - alt_text
      - credit
      - license_code
      - image
      title: Body_process_editorial_media_api_v1_admin_media_process_post
...
    CleanupRecoveryEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/CleanupRecoverySchema'
      type: object
      required:
      - data
      title: CleanupRecoveryEnvelope
    CleanupRecoveryRequest:
      properties:
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          default: 50
      type: object
      title: CleanupRecoveryRequest
    CleanupRecoverySchema:
      properties:
        completed:
          type: integer
          title: Completed
        failed:
          type: integer
          title: Failed
      type: object
      required:
      - completed
      - failed
      title: CleanupRecoverySchema
...
    EditorialMediaEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/MediaAssetRead'
      type: object
      required:
      - data
      title: EditorialMediaEnvelope
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          example: NOT_FOUND
        message:
          type: string
          title: Message
          example: O recurso solicitado não foi encontrado.
        details:
          anyOf:
          - {}
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: Detailed information about an API error.
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          title: Request Id
          example: req_123456789
      type: object
      required:
      - error
      - request_id
      title: ErrorResponse
      description: Standard error response envelope.
...
    GooglePhotoAttributionSchema:
      properties:
        display_name:
          type: string
          title: Display Name
        uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Uri
      type: object
      required:
      - display_name
      title: GooglePhotoAttributionSchema
    GooglePhotoMetadataEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/GooglePhotoMetadataSchema'
      type: object
      required:
      - data
      title: GooglePhotoMetadataEnvelope
    GooglePhotoMetadataSchema:
      properties:
        proxy_url:
          type: string
          title: Proxy Url
        expires_at:
          type: integer
          title: Expires At
        width_px:
          type: integer
          title: Width Px
        height_px:
          type: integer
          title: Height Px
        author_attributions:
          items:
            $ref: '#/components/schemas/GooglePhotoAttributionSchema'
          type: array
          title: Author Attributions
        google_maps_uri:
          type: string
          title: Google Maps Uri
      type: object
      required:
      - proxy_url
      - expires_at
      - width_px
      - height_px
      - google_maps_uri
      title: GooglePhotoMetadataSchema
      description: Ephemeral photo metadata; deliberately excludes Google resource/photo
        URLs.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MediaAssetRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        owner_type:
          type: string
          title: Owner Type
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        storage_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Storage Key
        mime_type:
          type: string
          title: Mime Type
        alt_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Alt Text
        credit:
          anyOf:
          - type: string
          - type: 'null'
          title: Credit
        license_code:
          anyOf:
          - type: string
          - type: 'null'
          title: License Code
        processing_status:
          type: string
          title: Processing Status
          default: pending
        checksum_sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Checksum Sha256
        width_px:
          anyOf:
          - type: integer
          - type: 'null'
          title: Width Px
        height_px:
          anyOf:
          - type: integer
          - type: 'null'
          title: Height Px
        derivatives:
          additionalProperties: true
          type: object
          title: Derivatives
        location:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Location
        processed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Processed At
        rejected_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejected Reason
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        sort_order:
          type: integer
          title: Sort Order
          default: 0
      type: object
      required:
      - id
      - owner_type
      - owner_id
      - mime_type
      title: MediaAssetRead
...
    ResolvedMediaItemSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        owner_type:
          type: string
          title: Owner Type
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        url:
          type: string
          title: Url
        derivatives:
          additionalProperties:
            type: string
          type: object
          title: Derivatives
        alt_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Alt Text
        credit:
          anyOf:
          - type: string
          - type: 'null'
          title: Credit
        license_code:
          anyOf:
          - type: string
          - type: 'null'
          title: License Code
        sort_order:
          type: integer
          title: Sort Order
          default: 0
      type: object
      required:
      - id
      - owner_type
      - owner_id
      - url
      title: ResolvedMediaItemSchema
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer